changed message format, longs to ints
Apparently longs are 8 bytes on my 64bit computer and 4 bytes on my 32bit
computer. I have changed the basic format of the proofnet message to
make this simpler... we will use unsigned ints for nonces and utc_times,
instead of unsigned longs. This means they are now 4 bytes instead of 8
bytes. There is probably plenty of room to last for many decades in the
utc_time... plus having some rollover is OK, the meaning of the utc_time
will just change if it ever rolls over. So 4 bytes is sufficient for
time. Hopefully it is also sufficient for the nonce... but if we ever
need more than 4 bytes for the 0s in the proof of work, then you may
need to add a nonce into the message as well.